refactor: splitting the migrations out to one file per version
